changing the reservation table
