+++ backend/alembic/versions/d3c3f0eb6761_reservation.py
@@ -1,0 +1,34 @@
+"""reservation
+
+Revision ID: d3c3f0eb6761
+Revises: f4ebe42757c8
+Create Date: 2025-01-28 20:32:34.355059
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'd3c3f0eb6761'
+down_revision: Union[str, None] = 'f4ebe42757c8'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('reservation', sa.Column('position', sa.Integer(), server_default=sa.text('0'), nullable=False))
+    op.add_column('reservation', sa.Column('is_satisfied', sa.Boolean(), server_default=sa.text('false'), nullable=False))
+    op.add_column('reservation', sa.Column('is_done', sa.Boolean(), server_default=sa.text('false'), nullable=False))
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('reservation', 'is_done')
+    op.drop_column('reservation', 'is_satisfied')
+    op.drop_column('reservation', 'position')
+    # ### end Alembic commands ###
