+++ backend/app/models/__init__.py
@@ -1,2 +1,3 @@
 from . import books
 from . import users
+from . import purchase
