+++ backend/app/models/books.py
@@ -1,8 +1,7 @@
 from app.database import Base
 from sqlalchemy.orm import relationship
-from sqlalchemy import Column, ForeignKey, String, DateTime, Integer
+from sqlalchemy import Column, ForeignKey, String, Integer
 
-from sqlalchemy.sql.functions import current_timestamp
 
 
 class Genre(Base):
@@ -37,12 +36,3 @@ class BookGenre(Base):
     book = relationship("Book")
 
 
-class Reservation(Base):
-    __tablename__ = "reservation"
-    client_id = Column(ForeignKey("client.id", ondelete="CASCADE"), nullable=False)
-    client = relationship("Client")
-    book_id = Column(ForeignKey("book.id", ondelete="CASCADE"), nullable=False)
-    book = relationship("Book")
-    start = Column("start", DateTime(), default=current_timestamp(), nullable=False)
-    end = Column("end", DateTime(), default=current_timestamp(), nullable=False)
-    price = Column(Integer, nullable=False)
